Revert "Merge branch 'main' of https://github.com/xyfx666/WiriScript_V4_CN"
This reverts commit 9f5d4c17d623c503c6a21631e487be980633846b, reversing
changes made to ffadd69f03443551823812617711e07e07c2e544.
