+++ README.md
@@ -2,10 +2,3 @@
 
 ## 一.我们是做什么的？  
 我们会翻译一些Stand的LUA脚本，并且会不定期添加项目，每个翻译的脚本都会在此仓库实时更新，并在翻译完成后发布Releases
-
-## 二.如何下载成品？
-如果您不想参加我们的翻译，只想要成品的话，您可以前往以下链接获取。  
-https://github.com/xyfx666/Stand-Script_CN/releases
-
-## 三.如何参加翻译？  
-您只需在您翻译完后提交PR即可。
